refactor bank client and add account id filtering to consent retrieval
